mi_agente_ai/requirements.txt
pydantic-ai
httpx
pydantic
python-dotenv
streamlit 